% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{get_input_attributes}
\alias{get_input_attributes}
\title{Get biomaRt input attributes for gene ID conversion}
\usage{
get_input_attributes(from_type)
}
\arguments{
\item{from_type}{Character string specifying the input gene ID type.}
}
\value{
A character string with the corresponding biomaRt attribute name.
}
\description{
This is a helper function that maps input gene ID types to the corresponding
biomaRt attribute names used for querying the Ensembl database.
}
\examples{
get_input_attributes("symbol")    # "external_gene_name"
get_input_attributes("ensembl")   # "ensembl_gene_id"
get_input_attributes("entrez")    # "entrezgene_id"
}
