% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_metadata_dir}
\alias{get_metadata_dir}
\title{Get path to the package metadata directory}
\usage{
get_metadata_dir()
}
\value{
Character string with the full path to the metadata directory.
}
\description{
Get path to the package metadata directory
}
