% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{get_output_attributes}
\alias{get_output_attributes}
\title{Get biomaRt output attributes for gene ID conversion}
\usage{
get_output_attributes(to_type)
}
\arguments{
\item{to_type}{Character string specifying the output gene ID type.}
}
\value{
A character string with the corresponding biomaRt attribute name.
}
\description{
This is a helper function that maps output gene ID types to the corresponding
biomaRt attribute names used for querying the Ensembl database.
}
\examples{
get_output_attributes("entrez")   # "entrezgene_id"
get_output_attributes("symbol")   # "external_gene_name"
get_output_attributes("ensembl")  # "ensembl_gene_id"
}
