% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_example_expression_data}
\alias{load_example_expression_data}
\title{Load example expression data}
\usage{
load_example_expression_data()
}
\value{
A data frame with expression data where rows are genes and columns are samples.
}
\description{
Loads the example expression data included with the tAge package.
}
\examples{
expr_data <- load_example_expression_data()
head(expr_data)
}
