% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_example_metadata}
\alias{load_example_metadata}
\title{Load example metadata}
\usage{
load_example_metadata()
}
\value{
A data frame with sample metadata.
}
\description{
Loads the example metadata included with the tAge package.
}
\examples{
meta_data <- load_example_metadata()
head(meta_data)
}
