% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_gene_list}
\alias{load_gene_list}
\title{Load gene list}
\usage{
load_gene_list()
}
\value{
A character vector of gene identifiers.
}
\description{
Loads the gene list included with the tAge package.
}
\examples{
gene_list <- load_gene_list()
head(gene_list)
}
