% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tAge-package.R
\docType{package}
\name{tAge-package}
\alias{tAge}
\alias{tAge-package}
\title{tAge: tAge -- Transcriptomic Biological Age Analysis Package}
\description{
tAge provides API for Gladyshev Lab transcriptomic clock use. The package implements preprocessing pipelines, age prediction using Elastic Net and Bayesian Ridge models, and comprehensive visualization tools for transcriptomic age analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Gladyshev-Lab/tAge}
  \item \url{https://gladyshev-lab.github.io/tAge/}
  \item Report bugs at \url{https://github.com/Gladyshev-Lab/tAge/issues}
}

}
\author{
\strong{Maintainer}: Dmitrii Glubokov \email{dglubokov@bwh.harvard.edu}

}
\keyword{internal}
