% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{map_genes}
\alias{map_genes}
\title{Map genes in an ExpressionSet using local CSV mapping tables}
\usage{
map_genes(eset, species, gene_mapping_type, verbose = TRUE)
}
\arguments{
\item{eset}{An ExpressionSet object.}

\item{species}{One of "human", "mouse", "rat", "monkey".}

\item{gene_mapping_type}{One of "Ensembl" or "Gene.Symbol".}

\item{verbose}{Logical. Print progress messages. Default TRUE.}
}
\value{
ExpressionSet with Entrez (mouse) gene IDs as rownames.
}
\description{
Map genes in an ExpressionSet using local CSV mapping tables
}
