% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_tAge}
\alias{predict_tAge}
\title{Predict transcriptomic age for multiple processed ExpressionSet objects}
\usage{
predict_tAge(tAge_eset, model_paths, species, mode)
}
\arguments{
\item{tAge_eset}{A named list of ExpressionSet objects, each representing a different
normalization method (e.g., "scaled", "scaled_diff", "yugene", "yugene_diff").}

\item{model_paths}{A named list of model paths corresponding to each normalization method.}

\item{species}{Character string specifying the species for the models.}

\item{mode}{Character string specifying the model type. Must be either "EN" for
Elastic Net or "BR" for Bayesian Ridge.}
}
\value{
A data frame containing the predicted transcriptomic age results for all
  provided ExpressionSet objects, with appropriately named columns.
}
\description{
This function predicts transcriptomic age for multiple processed ExpressionSet objects
using pre-trained models. It supports different normalization methods and model types.
}
